refactor: resolve every new-code SonarCloud issue on the #875 diff
The SonarCloud gate's 'fail on new issues' step blocks on all 95 smells on
lines this branch touched (moved/added lines count as new). Resolve them all at
source -- no suppressions, no gate changes:

- Split 7 files over the 500-line limit into cohesive sibling modules, kept as
  inherited mixins / re-exports so every public symbol stays importable and no
  caller changes (S104).
- Reduce cognitive/cyclomatic complexity and >3-return functions by extracting
  well-named, behavior-neutral helpers (S3776, FunctionComplexity, S1142).
- Bundle over-long parameter lists, add missing type hints/args, drop unused
  params, define constants for duplicated literals, fix trailing comments and
  static methods (S107, S6540, S6543, S1172, S1192, S139, S2325, S1128, S1720).
- Tests: split composite assertions, fix actual/expected order, narrow
  pytest.raises scopes (S9073, S3415, S5778).

Behavior preserved end to end: full non-integration suite 4368 passed, and a
full env-pack boot still reaches exit 0 with 0 SEM-218 diagnostics and 32
healthy containers.
