+++ src/aptl/backends/raes_content_realization.py
@@ -37,14 +37,26 @@ from __future__ import annotations
 import re
 from collections.abc import Mapping
 from dataclasses import dataclass
-from pathlib import Path, PurePosixPath
+from pathlib import Path
 from typing import Any
 
 from raes_contracts.diagnostics import Diagnostic
 from raes_contracts.planning import PlannedResource
 
+from aptl.backends._raes_content_spec import (
+    # Re-exported so the shared spec primitives keep their historical
+    # ``raes_content_realization`` names after the split.
+    _SAFE_DEST_RELPATH as _SAFE_DEST_RELPATH,
+    _is_sensitive,
+    _reject,
+    _safe_dest_relpath,
+)
+from aptl.backends._raes_dataset_content import (
+    _dataset_item_names as _dataset_item_names,
+    _dataset_tags as _dataset_tags,
+    _resolve_dataset_content,
+)
 from aptl.backends.raes_content_source_policy import forbidden_source_reason
-from aptl.backends.raes_diagnostics import diagnostic
 from aptl.backends.raes_realization_model import ParticipantDatasetRealization
 from aptl.backends.raes_realization_values import (
     content_source_exact_artifact as _content_source_exact_artifact,
@@ -69,7 +81,6 @@ _CONTENT_REALIZABLE_SERVICES: dict[str, str] = {
 }
 
 _RUNTIME_OBSERVED_PREFIX = "runtime-observed:"
-_SAFE_DEST_RELPATH = re.compile(r"^[A-Za-z0-9][A-Za-z0-9._/-]*$")
 
 
 @dataclass(frozen=True)
@@ -181,68 +192,6 @@ def _content_placement_inputs(
                 volume_suffix=volume_suffix,
             )
     return inputs, reason
-
-
-def _resolve_dataset_content(
-    *,
-    resource: PlannedResource,
-    spec: Mapping[str, Any],
-    content_name: str,
-    target_address: str,
-) -> tuple[ParticipantDatasetRealization | None, list[Diagnostic]]:
-    """Lower one empty logical evidence dataset into the run-store carrier."""
-
-    dataset: ParticipantDatasetRealization | None = None
-    diagnostics: list[Diagnostic] = []
-    forbidden_values = (
-        spec.get("source"),
-        spec.get("text"),
-        spec.get("path"),
-        spec.get("destination"),
-    )
-    if any(value not in (None, "") for value in forbidden_values):
-        diagnostics = [_reject(resource.address, "dataset-carries-materialization")]
-    else:
-        item_names = _dataset_item_names(spec.get("items"))
-        tags = _dataset_tags(spec.get("tags", []))
-        if item_names is None:
-            diagnostics = [_reject(resource.address, "dataset-items-invalid")]
-        elif tags is None:
-            diagnostics = [_reject(resource.address, "dataset-tags-invalid")]
-        else:
-            dataset = ParticipantDatasetRealization(
-                address=resource.address,
-                target_address=target_address,
-                content_name=content_name,
-                item_names=item_names,
-                tags=tags,
-                sensitive=_is_sensitive(spec),
-            )
-    return dataset, diagnostics
-
-
-def _dataset_item_names(raw_items: object) -> tuple[str, ...] | None:
-    """Validate and normalize the declared dataset item names."""
-
-    if not isinstance(raw_items, list) or not raw_items:
-        return None
-    item_names: list[str] = []
-    for item in raw_items:
-        name = item.get("name") if isinstance(item, Mapping) else None
-        if not isinstance(name, str) or not name or name in item_names:
-            return None
-        item_names.append(name)
-    return tuple(item_names)
-
-
-def _dataset_tags(raw_tags: object) -> tuple[str, ...] | None:
-    """Validate and normalize the declared dataset tags."""
-
-    if not isinstance(raw_tags, list) or any(
-        not isinstance(tag, str) or not tag for tag in raw_tags
-    ):
-        return None
-    return tuple(raw_tags)
 
 
 def _resolve_file_content(
@@ -442,12 +391,14 @@ def _resolve_pack_artifact_content(
     artifact_id = _optional_string(exact, "artifact_id")
     digest = _optional_string(exact, "digest")
     media_type = _optional_string(exact, "media_type")
-    if not artifact_id or not digest:
-        return None, [_reject(resource.address, "pack-artifact-missing-identity")]
-    if not _SHA256_DIGEST_RE.match(digest):
-        return None, [_reject(resource.address, "pack-artifact-invalid-digest")]
-    if source_kind == "pack-directory" and media_type != "application/x-tar":
-        return None, [_reject(resource.address, "pack-directory-not-archive")]
+    reason = _pack_artifact_reject_reason(
+        artifact_id=artifact_id,
+        digest=digest,
+        media_type=media_type,
+        source_kind=source_kind,
+    )
+    if reason is not None:
+        return None, [_reject(resource.address, reason)]
     content = DeploymentContentRealization(
         address=resource.address,
         target_address=placement.target_address,
@@ -463,6 +414,31 @@ def _resolve_pack_artifact_content(
     return content, []
 
 
+def _pack_artifact_reject_reason(
+    *,
+    artifact_id: str | None,
+    digest: str | None,
+    media_type: str | None,
+    source_kind: str,
+) -> str | None:
+    """Return the rejection reason for a malformed pack identity, else None.
+
+    Checked in the order the identity is established: an artifact id + digest
+    must be declared, the digest must be a canonical sha256, and directory
+    content must arrive as a tar archive.
+    """
+
+    if not artifact_id or not digest:
+        reason = "pack-artifact-missing-identity"
+    elif not _SHA256_DIGEST_RE.match(digest):
+        reason = "pack-artifact-invalid-digest"
+    elif source_kind == "pack-directory" and media_type != "application/x-tar":
+        reason = "pack-directory-not-archive"
+    else:
+        reason = None
+    return reason
+
+
 def _resolve_project_source(
     address: str, source_name: str, project_dir: Path
 ) -> tuple[Path | None, list[Diagnostic]]:
@@ -476,32 +452,3 @@ def _resolve_project_source(
     except PathContainmentError:
         return None, [_reject(address, "source-path-escapes-project")]
     return resolved, []
-
-
-def _safe_dest_relpath(relpath: str) -> bool:
-    """Return whether a destination relpath is safe to embed in a seed script."""
-
-    return (
-        ".." not in PurePosixPath(relpath).parts
-        and _SAFE_DEST_RELPATH.match(relpath) is not None
-    )
-
-
-def _is_sensitive(spec: Mapping[str, Any]) -> bool:
-    """Return the Content spec's `sensitive` flag as a plain bool."""
-
-    value = spec.get("sensitive")
-    return bool(value) if isinstance(value, (bool, str)) else False
-
-
-def _reject(address: str, reason_code: str) -> Diagnostic:
-    """Build a fail-closed content realization diagnostic."""
-
-    return diagnostic(
-        "aptl.provisioner.content-placement-rejected",
-        address,
-        (
-            "RAES content placement was rejected by the APTL content "
-            f"realization policy (reason={reason_code})."
-        ),
-    )
