+++ src/aptl/core/content_seed.py
@@ -23,7 +23,7 @@ import io
 import re
 import shutil
 import tarfile
-from collections.abc import Sequence
+from collections.abc import Callable, Sequence
 from pathlib import Path, PurePosixPath
 
 from aptl.core.credentials import (
@@ -65,15 +65,18 @@ def _seed_for_content(
     project_dir: Path, item: DeploymentContentRealization
 ) -> NamedVolumeSeed:
     """Build the named-volume seed for one typed content realization."""
-    if item.source_kind == "inline-text":
-        return _inline_text_seed(project_dir, item)
-    if item.source_kind in ("pack-file", "pack-directory"):
-        return _pack_artifact_seed(project_dir, item)
-    if item.source_kind in ("project-file", "project-directory"):
-        return _project_source_seed(project_dir, item)
-    # "empty-directory": an explicit empty-directory declaration has
-    # nothing to copy. The seed still runs (a harmless no-op `set -e`
-    # script) so the operation stays uniform for every content kind.
+    builder = _SEED_BUILDERS.get(item.source_kind, _empty_directory_seed)
+    return builder(project_dir, item)
+
+
+def _empty_directory_seed(
+    project_dir: Path, item: DeploymentContentRealization
+) -> NamedVolumeSeed:
+    """Seed an explicit empty-directory declaration, which has nothing to copy.
+
+    The seed still runs (a harmless no-op `set -e` script) so the operation
+    stays uniform for every content kind.
+    """
     return NamedVolumeSeed(
         volume_suffix=item.volume_suffix,
         source_dir=project_dir,
@@ -178,3 +181,16 @@ def _project_source_seed(
 def _content_slug(address: str) -> Path:
     """Return a filesystem-safe, code-defined subdirectory name for an address."""
     return Path(_SLUG_RE.sub("_", address))
+
+
+# The realization's declared source kind decides how its bytes reach the volume.
+# Any other kind (an explicit empty directory) seeds nothing.
+_SEED_BUILDERS: dict[
+    str, Callable[[Path, DeploymentContentRealization], NamedVolumeSeed]
+] = {
+    "inline-text": _inline_text_seed,
+    "pack-file": _pack_artifact_seed,
+    "pack-directory": _pack_artifact_seed,
+    "project-file": _project_source_seed,
+    "project-directory": _project_source_seed,
+}
